test(factory-sessions): isolate terminal response witnesses
